+++ HISTORY.rst
@@ -9,4 +9,5 @@ History
 0.2.0 (WIP)
 -----------
 
-* Add `.axis_indices`.
+* Add `.axis_indices` and `.axis_index`.
+* Add ``axis=value`` to names to slice 0D/1D of 1D/2D data.